requirements_dev.txt
# Requirements to be installed by developers
Django==4.0.4
djangorestframework==3.13.1
django-environ==0.8.1
django-cors-headers==3.11.0
django-simple-history==3.0.0
django-storages==1.12.3
django-extensions==3.1.5
boto3==1.22.1
requests==2.27.1
arrow==1.2.2
ipython==8.3.0